+++ src/backend/lib/async-service.d.ts
@@ -3,7 +3,6 @@ import {
   ListenerSignature,
   DefaultListener,
 } from "tiny-typed-emitter";
-import { AsyncServiceState } from "./types";
 
 // This is done this way in a d.ts file with the same name as `async-service.js`
 // since it is not possible to declare a class with Generics in JSDoc, so by
@@ -29,4 +28,20 @@ declare abstract class AsyncService<
   abstract _stop(): Promise<void>;
 }
 
-export = AsyncService;
+export type AsyncServiceStateValue =
+  | "stopped"
+  | "starting"
+  | "started"
+  | "stopping"
+  | "error";
+
+export type AsyncServiceState =
+  | {
+      value: Exclude<AsyncServiceStateValue, "error">;
+    }
+  | {
+      value: "error";
+      error: Error;
+    };
+
+export default AsyncService;
