+++ src/backend/lib/async-service.js
@@ -1,7 +1,7 @@
 // @ts-check
 const { TypedEmitter } = require("tiny-typed-emitter");
 
-/** @typedef {import('./types').AsyncServiceState } AsyncServiceState */
+/** @typedef {import('./async-service').AsyncServiceState } AsyncServiceState */
 
 /**
  * @typedef {Object} InternalEvents
